trending tags video feed created using getTrendingTagFeed gql query
